PLAT-7863: If new file download attempt failed unlink local file created

+++ batch/batches/Import/KAsyncImport.class.php
@@ -153,7 +153,7 @@ class KAsyncImport extends KJobHandlerWorker
 			
 			if($responseStatusCode && KCurlHeaderResponse::isError($responseStatusCode))
 			{
-				if(file_exists($data->destFileLocalPath))
+				if(!$resumeOffset && $data->destFileLocalPath && file_exists($data->destFileLocalPath))
 					unlink($data->destFileLocalPath);
 				
 				$this->closeJob($job, KalturaBatchJobErrorTypes::HTTP, KalturaBatchJobAppErrors::REMOTE_DOWNLOAD_FAILED, "Failed while reading file. HTTP Error: [$responseStatusCode]", KalturaBatchJobStatus::RETRY);
